% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amdemo-package.R
\docType{package}
\name{amdemo-package}
\alias{amdemo-package}
\title{amdemo: Runs Allelematch Test Driven Design}
\description{
Demostrates new functionality of package 'allelematch'. Currently the new parameter 'minComparableLoci'.
}
\author{
\strong{Maintainer}: Torvald Staxler \email{torvald.staxler@telia.com}

Other contributors:
\itemize{
  \item Department of Wildlife, Fish and Environmental Studies at Swedish University of Agricultural Sciences [copyright holder]
  \item Göran Spong \email{Goran.Spong@slu.se} (Senior Lecturer at the Department of Wildlife, Fish and Environmental Studies) [copyright holder]
  \item Paul Galpern \email{pgalpern@gmail.com} (Author of the excellent package 'allelematch' from which the five amExample data files have been copied under MIT license) [contributor]
}

}
\keyword{internal}
