% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amdemo-package.R
\name{amdemo}
\alias{amdemo}
\title{Package Overview}
\description{
Package 'amdemo' demonstrates new functionality of package \link{allelematch}.

The current version of this package demonstrates the use of
the new parmeter minComparableLoci.

The API is simple. There are only four functions:
\tabular{clcl}{
\verb{  } \tab \link{demoRun}\verb{  }\tab \verb{  } \tab Executes the demo tests, or a subset of the tests\cr
\verb{  } \tab \link{demoList}\tab \tab Lists the available demo tests without running them\cr
\verb{  } \tab \link{demoVersion}\tab \tab Shows the used versions of \link{allelematch} and \link{amdemo}\cr
\verb{  } \tab \link{demoInstallCranAllelematch}\tab\verb{  } \tab Installs official CRAN version of \link{allelematch}\cr
}
}
\references{
\link{amdemo-package}

\url{https://github.com/cran/allelematch}

\href{https://github.com/cran/allelematch/blob/2.5.1/inst/doc/allelematchSuppDoc.pdf}{allelematchSuppDoc.pdf}
}
