% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_api.R
\name{demoInstallCranAllelematch}
\alias{demoInstallCranAllelematch}
\title{Installs one of the official versions of 'allelematch' from CRAN.}
\usage{
demoInstallCranAllelematch(version = "2.5.4")
}
\arguments{
\item{version.}{Default "2.5.4".}
}
\value{
TBD
}
\description{
TBD\cr
\cr
TBD
}
\examples{
# Install the default official version of 'allelematch' from CRAN:
ttInstallCranAllelematch()

# Install another official version of 'allelematch' from CRAN:
ttInstallCranAllelematch("2.5.3")


}
\seealso{
\link{demoVersion} and \link{amamdemo}
}
