% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_api.R
\name{demoList}
\alias{demoList}
\title{Lists available tests in \code{amdemo} without running them}
\usage{
demoList(verbose = TRUE)
}
\arguments{
\item{verbose}{logical. If TRUE (the default), prints additional info to stdout}
}
\value{
A character vector containing the names of all the tests
}
\description{
Use the output to select a value for parameter \code{filter} to \link{demoRun}.
Useful when debugging.
}
\examples{
# See what version of packages 'allelematch' and 'amdemo'
# are currently installed:
demoVersion()

# List the available tests:
demoList()
\donttest{
# Run all the tests:
# demoRun()  # Takes several minutes

# Run the first of the available tests:
demoRun(filter="allelematch_1-amDataset$")
}

}
\seealso{
\link{demoVersion} and \link{demoRun}
}
