% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_api.R
\name{demoRun}
\alias{demoRun}
\title{Runs the regression test}
\usage{
demoRun(filter = "", verbose = TRUE)
}
\arguments{
\item{filter}{If specified, only tests with names matching this perl regular
expression will be executed. Character vector of length 1. See also \link{demoList}}

\item{verbose}{logical. If TRUE (the default), prints version of tested allelematch to stdout}
}
\value{
A list (invisibly) containing data about the test results as returned by \link[testthat:test_package]{testthat::test_package}
}
\description{
Runs regression tests on package \link{allelematch} to make sure it is backwards compatible.\cr
\cr
The full set of tests will take a couple of minutes. \cr
\cr
Call \link{demoList} to see the available tests with without running them.
}
\details{
If any of the test executed with \link{demoRun} should fail, then we want to be able
to run that specific test under the debugger.\cr
\cr
Set a breakpoint in \code{allelematch.R} and call \code{demoRun(filter="<the test that reproduces the problem>")}\cr
\cr
Note that it is the last installed version of \code{allelematch} that will be executed,
not the last edited. In RStudio, CTRL+SHIFT+B will build and install.
}
\examples{
# See what version of packages 'allelematch' and 'amdemo'
# are currently installed:
demoVersion()

# List the available tests:
demoList()
\donttest{
# Run all the tests:
# demoRun()  # Takes several minutes

# Run the first of the available tests:
demoRun(filter="allelematch_1-amDataset$")
}

}
\seealso{
\link{demoVersion} and \link{demoList}
}
