% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_api.R
\name{demoVersion}
\alias{demoVersion}
\title{Returns package version}
\usage{
demoVersion(packages = c("amdemo", "amregtest", "allelematch"), verbose = TRUE)
}
\arguments{
\item{verbose}{logical. If TRUE (the default), prints additional info to stdout, including version of \link{allelematch-package}}
}
\value{
The installed version of this package (\link{amdemo-package}) in a character vector of length one
}
\description{
Displays version of this package (\link{amdemo}) and of \link{allelematch},
together with build timestamps.\cr
\cr
The version is specified in the file DESCRIPTION, tag "Version: ".\cr
\cr
Note that installing a package from source (like e.g. \link{allelematch}
it will show a build timestamp from the installation, not from when
it was published.

Returns version of this package (\link{amdemo}).\cr
\cr
The version is specified in the file DESCRIPTION, tag "Version: ".
}
\examples{
# See what version of packages 'allelematch' and 'amdemo'
# are currently installed:
demoVersion()

# List the available demo tests:
demoList()
\donttest{
# Run all the demo tests:
# demoRun()  # Takes several minutes

# Run the first of the available demo testss:
demoRun(filter="allelematch_1-amDataset$")
}


}
\seealso{
\link{demoList}, \link{demoRun} and \link{amdemo}
}
